Rework server pipelines and host calls (#7)
* Rework server pipelines and host calls

Restructure how pipelines are defined to better suite the IPC model used
with OutOfProc transports. Adds helper class for the client and server
to request and respond to host calls with valid data. Added integration
tests for both a client and server endpoint using a rough OutOfProc
transport. Various other bugfixes that came up when writing the
integration tests.