Do not use ref for lists and dicts
Avoid using a ref during a list/dict serialization. There is no
guarantee that the id() for these types will be unique during the
serialization process as certain property types and classes could
generate a short lived object causing conflicts in the id() value. The
values of the list can still be created as a reference placeholder if
they contain a complex type that is referencable.