Fix serialization of broken surrogates in strings
Fix the serialization of strings that contain broken surrogate pair
chars. This is required as .NET strings are essentially an array of char
objects. Surrogate pairs are composed of 2 chars which means a string
can contain half a surrogate pair.

This also fixes up the serialization of other string like types to
handle such scenarios like this.